Changed the child order in the suffix tree back and updated the iterator code.
It is easier to traverse in the right order in other algorithms, and it isn't much harder to deal with the reverse order in the iterators. When you use iterators, there is more overhead anyway, so it won't cost much there compared to dealing with the order in direct traversal algorithms.